+++ lib/view/list_page.dart
@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:intl/intl.dart';
-import 'package:simple_todo_with_latest_widget/model/list_page_riverpod.dart';
-import 'package:simple_todo_with_latest_widget/model/common_model.dart';
+import 'package:simple_todo_with_latest_widget/view_model/common_model.dart';
+import 'package:simple_todo_with_latest_widget/view_model/riverpod/list_page_riverpod.dart';
 
 class ListPage extends HookConsumerWidget {
   const ListPage({super.key});
