+++ lib/view/login_page.dart
@@ -4,9 +4,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:simple_todo_with_latest_widget/model/common_model.dart';
 import 'package:simple_todo_with_latest_widget/model/field.dart';
-import 'package:simple_todo_with_latest_widget/view_model/login_page_view_model.dart';
+import 'package:simple_todo_with_latest_widget/view_model/auth.dart';
+import 'package:simple_todo_with_latest_widget/view_model/common_model.dart';
+import 'package:simple_todo_with_latest_widget/view_model/router.dart';
 
 class LoginPage extends HookWidget {
   const LoginPage({super.key});
